Move OSD update loop into OSDAggregator.
It should've been in here all along.

This model simplifies things a lot. The background of the progress bar
has been split into its own class which does not update in the regular
event cycle—only when the size changes.

Additionally, the ASS for the display is more intelligent. Using \an1
instead of \an7 allows the position of the lines to simply be
width,height rather than width,(height - bar_height).

The bar itself is now the width of the video window and only a single
fscx value is updated each render loop. This saves a bunch of math,
because the fscx value is simply equal to the pos-percent property.