Main: fix initial draw.
Previously, the progress bar was not being drawn when mpv was launched
unless fullscreen was toggled.

It turns out that the `fullscreen` property observer in OSDAggregator
fires as soon as mpv starts. Since our ProgressBar and friends no longer
need dimensions passed to their constructor, they should be initialized
immediately, so that they are already attached as subscribers to the
aggregator when the observer fires.