Add animation queue and module.
Big changes afoot. This probably should have been lumped into a single
commit.

Reintroduces the Rect class. Used for determining if the mouse is
hovering a particular Rect within the window. ProgressBar and
ProgressBarBackground inherit from it.

OSDAggregator passes the current mouse position to its subscribers'
update callback, for use with hover effects.

The animation class takes a callback and uses it to animate a value
across a duration. It must be attached to an instance of an
AnimationQueue, which updates an aggregator when it has active
animations. The AnimationQueue repeatedly calls the update method of
each animation registered to it, as well as the forceUpdate method of
the aggregator associated with it.