OSDAggregator: add removeSubscriber method.
All subscribers have the field aggregatorIndex attached to them when
they subscribe. This can be used to unsubscribe from the index at any
time.

I thought about implementing this with a doubly linked list, the same
way AnimationQueue is implemented, but it's more trouble than it's
worth. Presumably an aggregator will never have enough subscribers to
make the advantages of a doubly linked list actually apparent.