Add bounds calculation with SubInspector to various UI elements.
This code will probably never be merged into master, as it requires
bundling a C library along with the script, which makes cross-platform
support an absolute nightmare. It was a fun experiment that has been
sitting around on my hard drive for the last half a year or so, and it
works surprisingly well. I'd rather have it here than lost forever,
which is why I'm committing it.

Also includes a timer for measuring performance, which is not currently
hooked up to anything in particular.