Add janky workaround to try to improve mouse behavior.
Attempts to fix #10. Since the hitboxes change when fullscreen is
toggled and mpv still has some issues with properly updating the mouse
position and seems to leak single click events, it was possible to
accidentally instantaneously seek when double clicking to toggle full
screen. Try to mitigate this by killing the mouse on fullscreen cycle,
until it has been moved. This may not help in the case that the mouse is
moved a single pixel as the transition happens. Additionally, the seek
command was occasionally firing before the fullscreen toggle
notification came in. Wrapping the seek command in a timer seems to have
helped with this but is a suboptimal solution.

Overall, I'm not a huge fan of this commit, though it does, on cursory
inspection, seem to fix what it set out to.