Subscriber: don't inherit from Rect.
Subscriber now holds an instance of the hover rect (@zone) used for OSD
activation. This doesn't change any functionality, but I think it makes
more sense semantically because it avoids conflating the idea of the
actual bounding rect of a subscriber and the hover activity zone.

Some subscriber subclasses still use this as an alias for window width
for positioning, which is something that should be fixed in the future.