ProgressBarCache: make cache always appear ahead of position.
Previously, the cache was added to a dumb guess of current play position
based on a naïve average bitrate calculation. While the amount of video
cached is still based on an ABR calculation, but at least it is now
appended to the actual position, which keeps the cache bar from actually
falling behind the progress bar, a problem that it had before.