ActivityZone: handle de/activation details.
The main point of introducing the ActivityZone class was to offload the
redundant hit checks occurring in every UI element into a more
consolidated location. This turns out to have been more complicated
than initially anticipated.

It's worth noting that the original OSDAggregator code went to great
pains to avoid looping over the list of UI elements more than once. With
the recent reorganization (including the introduction of ActivityZone
and the removal of the second timer from AnimationQueue) this no longer
appears to be possible.