Settings: switch UI styling to ASS override strings.
While this approach is almost certainly less immediately user friendly,
the significant gain in configurability has been deemed to outweigh the
detriments. It also offloads handling style "inheritance", as it were,
to libass, removing the kludgy placeholder string technique previously
used for color inheritance.

This will result in longer strings being passed around to the renderer,
something that was previously strenuously avoided (and part of the
reason this solution was not adopted until this point). However, the
enhanced configurability and code simplicity was also deemed to outweigh
whatever entirely negligible performance hit occurs due to this change.

This is also, to my knowledge, the first backwards-incompatible change
that has been introduced to the settings. It seems to me that
implementing and testing a backwards-compatibility shim would be far too
much trouble, so such a shim has been omitted.