Main: fix PauseIndicator crash under certain startup conditions.
In particular, certain scenarios would immediately cause an instance of
PauseIndicator to be constructed, adding it to both eventLoop.uiElements
and the AnimationQueue. Then, when eventLoop\generateUIFromZones was
called later in the script initialization, eventLoop.uiElements was
cleared, but AnimationQueue wasn't, causing the PauseIndicator animation
destructor to try to remove itself from the already cleared uiElements.

This fixes the problem by also clearing out the AnimationQueue when
creating the UI. I don't think the user wants a PauseIndicator to appear
when mpv first launches anyway.

Fixes #37.