HoverTime: animate position from bottom instead of alpha.
I liked the fade-in effect, but it had the problem that it would
override any style-configured opacity for the hover time. Making it pop
up from behind the progress bar is a significantly more straightforward
solution than trying to parse the configured line style for opacity
tags, so the animation has been changed. I think this will work better
with the (eventually) planned chapter title hover effect.