all: declare instance variables in class body.
This is the result of a bit of a mess, and I will probably regret it.
All classes should call reconfigure on construction, and all classes
inherit from the base class UIElement, so it makes sense to call
reconfigure from the UIElement constructor. However, I also think that
classes should call their superclass constructor at the beginning of
their constructor.

A consequence of this was that reconfigure would end up being run before
the constructor and attempt to access instance variables that don't
exist. A mediocre, although acceptable solution, would be to simply move
the contents of the constructor to reconfigure. By and large, the two
methods were essentially duplicates anyway. However, while the
constructor is only called once, reconfigure can be called many times,
so certain operations didn't make sense (such as constructing a fresh
line table)

As it turns out, moonscript allows declaring ivars in the class body,
which seems like the best solution, except that the actual
implementation has some quirks. Like methods, these ivars are actually
stored in the instance's metatable. This means that they are shared
between all instances of the class until first assignment. For table
variables, which are reference types, mutating them will mutate the
ivar for all instances. Currently, this is not a problem because all of
the UIElement subclasses are intended to be (and currently used as)
singletons, but this is not a particularly robust solution.

Another (annoying) side-effect of moving the ivar declarations into the
class body is that they are assigned before anything else, so their
declarations cannot reference class-local variables, class variables, or
other instance variables.