main: simplify initial draw callback chain.
I don't remember why this code listened for two events, but at some
point this started behaving nondeterministically. It seems that the
`file-loaded` event is perfectly adequate for launching the initial
draw with recent-ish mpv, based on a comprehensive test suite of loading
a single file, once. This is the kind of high-quality development effort
that open source software can get you.