EventLoop: avoid unnecessary redraws.
Apparently mpv 0.31 changed the redrawing logic in such a way that
posting duplicate redraws (i.e. a draw that is the same as the previous
draw) becomes quite CPU intensive. This is most obviously noticeable on
pause, when duplicate redraws are much more likely to be submitted
(during playback, the percent-pos property changes pretty much every
frame, making duplicate redraws very unlikely).

This commit also adds some basic caching logic to the ProgressBarCache
element, which, perhaps ironically, was never being cached before. This
is mostly an all-or-nothing brute-force caching approach, but on
initial inspection it appears to actually work.

In addition, ProgressBar and ProgressBar cache no longer submit redraw
requests if they are set to hide-inactive and are actually inactive.
This is a smaller optimization, but it conforms to the expected
behavior, increasing consistency.

Fixes: torque/mpv-progressbar#56