HoverTime: redraw when duration changes
This was a bit of an oversight on how this was redrawn previously. For
example changing between two videos with the same resolution but
different duration, the time hovered would not update until the mouse
moved.

In theory it's possible for the duration to change while a video is
playing I guess (but normally this would be for streamed video, for
which we don't present the hover time at all) so that's another edge
case potentially handled.