HoverTime: fix redraw problem across unknown duration
When duration went to 0 and then came back with the same value it had
previously (e.g. playing the same video twice in a row), the time
would change to ???? and not change back to a timestamp until the
mouse was moved. This happened because the computed hover timestamp
did not get updated during the portion of unknown duration, so the
value was cached too aggressively.

By decoupling the positioning logic from the timestamp logic, the
intent becomes more obvious and hopefully less bug-prone.