ProgressBarCache: fix full bar drawing
This is a hack that injects a newline to force this single component to
produce two separate lines for rendering. in order to produce smaller
output, the previous implementation baked both drawings into a single
line using an interest hack with the way that ASS drawings work. Well,
it seems that interesting hack was more with the way that libass
works, and at some point the libass behavior changed, which broke the
rendering.

When a video was fully cached, the bar would become 100% the width of
the canvas. This would cause libass's line-wrapping behavior to
engage, which would result in the bar being drawn incorrectly. This
was fixed by splitting the bar into two separate. Positioned
independently, they do not collide with each other and there is no
problem.