Chapters: add functionality to seek to the nearest chapter on click
This is not implemented particularly gracefully. Currently, it must be
bound to a separate button than standard click to seek, and it will
always seek to a chapter, even if the mouse is not particularly close
to a chapter marker.

It might make sense to have a setting that seeks to a chapter if the
mouse is within a certain threshold of the chapter marker, but that's
more difficult to implement and also requires pixels-on-screen
computation for the hitbox, which is messy.

This feature is disabled by default for backward compatibility, since
the default keybindings have MBTN_RIGHT bound to cycle pause.