Thumbnail: add optional border box
This is now enabled by default. This does not have a specific setting
to disable it, but it can be made invisible by overriding its style or
setting thumbnail-border-expansion = 0, in which case the box will be
(mostly) covered by the thumbnail itself (it may be partially visible
when not aligned to the pixel grid, esp on high DPI displays).

The base layer is drawn as a box, and by default the border is
disabled, causing it to have sharp corners. Rounded corners can be
achieved by setting thumbnail-border-expansion = 0 and then
adding \bord5 to thumbnail-border-style. A two-color border (if one
wants a box with good contrast over both light and dark backgrounds)
can be achieved by having nonzero thumbnail-border-expansion and
setting nonzero border with a different border color in the style,
though the border will always have rounded corners.