Thumbnail: clamp border to within margins
This changes the logic so that the border expansion is incorporated
into the margins. This will have interesting results if
thumbnail-border-expansion is set to a negative value, and that's fine
by me (the correct way to hide it is to make it fully transparent).
Any border style added has to be manually accounted for using the
margin settings.