thumbnail: clear cached border drawing if thumbfast becomes disabled
For example, when playing a playlist that started with a local file and
then switched to a network file, the border box would stop updating,
but it would still be drawn.

Also change the defaults so it does not emit invalid ASS to start with,
though some of the other widgets may still do this.